pyproject.toml
[tool.poetry]
name = "leeky"
version = "0.1.0"
description = "leeky - training data contamination techniques for blackbox models"
authors = ["Michael Bommarito <michael.bommarito@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.61.0"
numpy = "^1.24.1"
spacy = "^3.7.2"
rapidfuzz = "^2.13.7"
httpx = {extras = ["h2"], version = "^0.23.3"}
rich = "^13.2.0"
playwright = "^1.30.0"
transformers = "^4.26.0"
accelerate = "^0.15.0"
sentencepiece = "^0.1.97"
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl"}
pyyaml = "^6.0.1"
matplotlib = "^3.8.2"
pandas = "^2.1.4"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
